Refactor import lookup builders
