+++ services/import_service_v2.py
@@ -74,6 +74,13 @@ from utils.serialization import (
     serialize_participant_event,
     serialize_participant,
 )
+from services.import.lookup_builders import (
+    DOC_TYPE_CACHE,
+    build_lookup_main_online,
+    build_lookup_participantslista,
+    finalize_doc_type_cache,
+)
+from services.import.normalize import normalize_text
 
 # ==============================================================================
 # 1. Configuration & Constants
@@ -88,9 +95,6 @@ if _participant_repo:
     initialize_cache(_participant_repo)
 
 PREVIEW_PARTICIPANT_LOOKUP = True
-
-_DOC_TYPE_CACHE: dict[str, str] = {}
-_DOC_TYPE_SEEN: set[str] = set()
 
 # ==============================================================================
 # 2. Custom XML Extraction and Parsing Utilities
@@ -399,145 +403,6 @@ def _load_custom_xml_objects(path: str) -> Optional[Dict[str, Any]]:
 # 8. Lookup Builders (ParticipantsLista / MAIN ONLINE)
 # ==============================================================================
 
-def _build_lookup_participantslista(df_positions: pd.DataFrame) -> Dict[str, Dict[str, str]]:
-    """
-    Build lookup from the 'ParticipantsLista' sheet.
-
-    Key:
-        'LAST|First Middle'
-    Value:
-        {
-            "position": ...,
-            "phone": ...,
-            "email": ...
-        }
-    """
-    name_col  = next((c for c in df_positions.columns if "name (" in c.lower()), None)
-    pos_col   = next((c for c in df_positions.columns if "position" in c.lower()), None)
-    phone_col = next((c for c in df_positions.columns if "phone" in c.lower()), None)
-    email_col = next((c for c in df_positions.columns if "email" in c.lower()), None)
-
-    look: Dict[str, Dict[str, str]] = {}
-    if not name_col:
-        return look
-
-    for _, row in df_positions.iterrows():
-        raw = _normalize(str(row.get(name_col, "")))
-        key = _name_key_from_raw(raw)
-        if not key:
-            continue
-        phone_value = normalize_phone(row.get(phone_col, "")) if phone_col else None
-        look[key] = {
-            "position": _normalize(str(row.get(pos_col, ""))) if pos_col else "",
-            "phone":    phone_value or "",
-            "email":    _normalize(str(row.get(email_col, ""))) if email_col else "",
-        }
-    return look
-
-
-def _build_lookup_main_online(df_online: pd.DataFrame) -> Dict[str, Dict[str, object]]:
-    """
-    Build lookup from the 'MAIN ONLINE → ParticipantsList' table.
-
-    Key:
-        'LAST|First Middle'  (plus fallback 'LAST|First')
-    Value:
-        normalized field dictionary with translated and enriched values.
-    """
-    cols = {c.lower().strip(): c for c in df_online.columns}
-
-    def col(label: str) -> Optional[str]:
-        return cols.get(label.lower())
-
-    look: Dict[str, Dict[str, object]] = {}
-    for _, row in df_online.iterrows():
-        first  = _normalize(str(row.get(col("Name")) or ""))
-        middle = _normalize(str(row.get(col("Middle name")) or ""))
-        last   = _normalize(str(row.get(col("Last name")) or ""))
-
-        if not first and not last:
-            continue
-
-        first_middle = " ".join(part for part in [first, middle] if part).strip()
-        key  = _name_key(last, first_middle)
-        keys = [key]
-        if middle and first:
-            keys.append(_name_key(last, first))  # Fallback
-
-        # --- Gender normalization ---
-        gender_col = col("Gender")
-        gender_raw = (str(row.get(gender_col, "")) if gender_col else "").strip()
-        normalized_gender = _normalize_gender(gender_raw)
-        gender = normalized_gender.value if normalized_gender else gender_raw
-
-        # --- Birth country translation ---
-        birth_country_raw  = re.sub(r",\s*world$", "", _normalize(str(row.get(col("Country of Birth"), ""))), flags=re.IGNORECASE)
-
-        # --- Travel document type ---
-        travel_doc_type_col = col("Traveling document type")
-
-        travel_doc_type_raw = _collect_doc_type(
-            row.get(travel_doc_type_col, "") if travel_doc_type_col else ""
-        )
-        # --- Transport and banking fields ---
-        transportation_col     = col("Transportation")
-        transport_other_col    = col("Transportation (Other)")
-        iban_type_col          = col("IBAN Type")
-
-        transportation_value   = str(row.get(transportation_col, "")) if transportation_col else ""
-        transport_other_value  = str(row.get(transport_other_col, "")) if transport_other_col else ""
-        iban_type_value        = str(row.get(iban_type_col, "")) if iban_type_col else ""
-
-        # --- Compose normalized entry ---
-        phone_col = col("Phone number")
-        phone_raw = row.get(phone_col, "") if phone_col else ""
-        phone_list_value = normalize_phone(phone_raw) or ""
-
-        entry = {
-            "name": _to_app_display_name(" ".join([first, middle, last]).strip()),
-            "gender": gender,
-            "dob": row.get(col("Date of Birth (DOB)")),
-            "pob": _normalize(str(row.get(col("Place Of Birth (POB)"), ""))),
-            "birth_country": birth_country_raw,
-            "citizenships": [
-                _normalize(x)
-                for x in re.split(r"[;,]", str(row.get(col("Citizenship(s)"), "")))
-                if _normalize(x)
-            ],
-            "email_list": _normalize(str(row.get(col("Email address"), ""))),
-            "phone_list": phone_list_value,
-            "travel_doc_type": travel_doc_type_raw,
-            "travel_doc_number": _normalize(str(row.get(col("Traveling document number"), ""))),
-            "travel_doc_issue": row.get(col("Traveling document issuance date")),
-            "travel_doc_expiry": row.get(col("Traveling document expiration date")),
-            "travel_doc_issued_by": translate(
-                _normalize(str(row.get(col("Traveling document issued by"), ""))), "en"
-            ),
-            "transportation_declared": transportation_value.strip(),
-            "transport_other": transport_other_value.strip(),
-            "traveling_from_declared": _normalize(str(row.get(col("Traveling from"), ""))),
-            "returning_to": _normalize(str(row.get(col("Returning to"), ""))),
-            "diet_restrictions": _normalize(str(row.get(col("Diet restrictions"), ""))),
-            "organization": translate(_normalize(str(row.get(col("Organization"), ""))), "en"),
-            "unit": translate(_normalize(str(row.get(col("Unit"), ""))), "en"),
-            "rank": translate(_normalize(str(row.get(col("Rank"), ""))), "en"),
-            "intl_authority": _normalize(str(row.get(col("Authority"), ""))),
-            "bio_short": translate(_normalize(str(row.get(col("Short professional biography"), ""))), "en"),
-            "bank_name": _normalize(str(row.get(col("Bank name"), ""))),
-            "iban": _normalize(str(row.get(col("IBAN"), ""))),
-            "iban_type": iban_type_value.strip(),
-            "swift": _normalize(str(row.get(col("SWIFT"), ""))),
-        }
-
-        for nk in keys:
-            if not nk:
-                continue
-            if nk not in look:
-                look[nk] = entry
-
-    return look
-
-
 # ==============================================================================
 # 9. Column Finder and Main Parsing Routine
 # ==============================================================================
@@ -624,10 +489,10 @@ def parse_for_commit(path: str, *, preview_only: bool = True) -> dict:
     df_positions = _read_table_df(path, plist, cache)
     df_online = _read_table_df(path, ponl, cache) if ponl else pd.DataFrame()
 
-    positions_lookup = _build_lookup_participantslista(df_positions)
-    online_lookup = _build_lookup_main_online(df_online) if not df_online.empty else {}
+    positions_lookup = build_lookup_participantslista(df_positions)
+    online_lookup = build_lookup_main_online(df_online) if not df_online.empty else {}
 
-    _finalize_doc_type_cache()
+    finalize_doc_type_cache()
 
     if DEBUG_PRINT:
         print(f"[STEP] Positions lookup entries: {len(positions_lookup)}")
@@ -886,30 +751,7 @@ def parse_for_commit(path: str, *, preview_only: bool = True) -> dict:
 
 def _normalize(s: Optional[str]) -> str:
     """Normalize whitespace and coerce None to an empty string."""
-    return re.sub(r"\s+", " ", (s or "").strip())
-
-def _collect_doc_type(value: object) -> str:
-    """Collect raw travel document values without normalizing yet."""
-    if not value:
-        return ""
-
-    raw = str(value).strip()
-    if raw:
-        _DOC_TYPE_SEEN.add(raw)
-    return raw
-
-def _finalize_doc_type_cache() -> None:
-    """Normalize all collected document types exactly once."""
-    for raw in _DOC_TYPE_SEEN:
-        key = re.sub(r"[^a-z0-9]+", " ", raw.lower()).strip()
-
-        # Passport detection
-        if "pass" in key:
-            normalized = str(DocType.passport.value)
-        else:
-            normalized = str(DocType.id_card.value)
-
-        _DOC_TYPE_CACHE[raw] = normalized
+    return normalize_text(s)
 
 
 # ==============================================================================
